fix: 분리 UX·파형 갱신·DB 저장 경쟁 등 개선 7건
- 다른 곡 분리 중에는 사유를 표시한 비활성 버튼으로 안내 (조용히 무시되던 것 수정)
- 앱 시작 프리캐치 완성 시 파형 자동 로드 (최대 60초 대기, 실패 감지 시 즉시 중단)
- SongDao 컬럼별 UPDATE로 저장 lost-update 제거, 미사용 Song.withStemLevels 제거
- PIPELINE_SAMPLE_RATE 상수화 — 불일치 샘플레이트 스템은 재생·내보내기에서 제외
- AudioDecode 스테레오 변환 버퍼 세션당 재사용 (출력 버퍼마다 재할당하던 것 개선)
- 라이브러리 캐시 준비 실패 Log.w, SpectralStage 불필요한 toFloat 제거
- AGENTS.md 불변식 갱신, README 구조·파형 설명 최신화

+++ app/src/main/java/com/bandmr/app/audio/SpectralStage.kt
@@ -271,7 +271,7 @@ class SpectralStage(private val sampleRate: Int, channels: Int = 2) {
             val hi = (center + 2.5f * sigma).toInt().coerceAtMost(n / 2)
             for (j in lo..hi) {
                 val d = j - center
-                val g = NOTCH_GAIN * kotlin.math.exp(-(d * d) / (2f * sigma * sigma)).toFloat()
+                val g = NOTCH_GAIN * kotlin.math.exp(-(d * d) / (2f * sigma * sigma))
                 val keep = 1f - g
                 re[j] *= keep; im[j] *= keep
                 if (j < n / 2) {
